Fix overflowing of indptr when slicing large datasets

+++ gemma-core/src/main/java/ubic/gemma/core/loader/expression/singleCell/AnnDataSingleCellDataLoader.java
@@ -274,7 +274,7 @@ public class AnnDataSingleCellDataLoader implements SingleCellDataLoader {
             log.warn( "Detecting counts from dense matrices are not supported, ignoring " + layer + "." );
             return false;
         }
-        for ( int i = 0; i < data.size(); i += 1000 ) {
+        for ( long i = 0; i < data.size(); i += 1000 ) {
             double[] vec = data.slice( i, Math.min( i + 1000, data.size() ) ).toDoubleVector();
             for ( double d : vec ) {
                 double uv;
